+++ Patches/ActionWheelSystemPatch.cs
@@ -1,6 +1,10 @@
 ﻿using HarmonyLib;
 using ProjectM.UI;
+using RetroCamera.Configuration;
 using RetroCamera.Utilities;
+using static RetroCamera.Configuration.QuipManager;
+using static RetroCamera.Utilities.Quips;
+using static RetroCamera.Systems.RetroCamera;
 
 namespace RetroCamera.Patches;
 
@@ -31,5 +35,50 @@ internal static class ActionWheelSystemPatch
             _wheelVisible = true;
             CameraState.IsMenuOpen = true;
         }
+
+        // Core.Log.LogWarning($"IsWheelActive {ActionWheelSystem.IsWheelActive}");
+    }
+
+    static DateTime _lastQuipSendTime = DateTime.MinValue;
+    const float QUIP_COOLDOWN_SECONDS = 1.5f;
+
+    [HarmonyPatch(typeof(ActionWheelSystem), nameof(ActionWheelSystem.SendQuipChatMessage))]
+    [HarmonyPrefix]
+    static bool SendQuipChatMessagePrefix(byte index)
+    {
+        DateTime now = DateTime.UtcNow;
+
+        if ((now - _lastQuipSendTime).TotalSeconds < QUIP_COOLDOWN_SECONDS)
+            return false;
+
+        _lastQuipSendTime = now;
+
+        if (CommandQuips.TryGetValue(index, out CommandQuip commandQuip))
+        {
+            SendCommandQuip(commandQuip);
+            return false;
+        }
+
+        return true;
+    }
+
+    [HarmonyPatch(typeof(ActionWheelSystem), nameof(ActionWheelSystem.HideCurrentWheel))]
+    [HarmonyPrefix]
+    static bool HideCurrentWheelPrefix(ActionWheelSystem __instance)
+    {
+        if (SocialWheelActive)
+        {
+            return false;
+        }
+
+        // Core.Log.LogWarning($"[ActionWheelSystem.HideCurrentWheel]");
+        return true;
+    }
+
+    [HarmonyPatch(typeof(ActionWheelSystem), nameof(ActionWheelSystem.UpdateAndShowWheel))]
+    [HarmonyPrefix]
+    static void UpdateAndShowWheelPrefix(ActionWheelSystem __instance)
+    {
+        // Core.Log.LogWarning($"[ActionWheelSystem.UpdateAndShowWheel]");
     }
 }