+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+`1.4.3`
+- added command wheel (right alt default key), must be enabled from menu and commands set in config file commands.json (first spot for name of command you want showing in wheel, second spot for raw command string)
+- can set size scaling of crosshair in menu
+- mouse should show when building during action mode again
+
 `1.3.2`
 - vertical aim offset appears to now be functioning as expected
 - option to hide character info panel at the top of the screen during action mode